requirements.txt
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==4.4.2
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.8
instagrapi==2.0.0
moviepy==1.0.3
numpy==1.25.2
oauthlib==3.2.2
Pillow==10.0.1
proglog==0.1.10
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
pydantic==1.10.9
pyparsing==3.1.1
PySocks==1.7.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4